improve tests checking that pooled BufferRecycler is properly acquired/released
